requirements.txt
transformers
PyYAML
torch
notebook
boto3
mypy-boto3-s3
python-box
ensure
-e .
